Moved a11ystring to TimeUtil
